feat : Review 기능 구현 (#70)
* add : resources 파일 추가

* feat: 판매 댓글 기능 구현중

* Refacor : comment 리팩터링

* refactor : comments 리팩터링

* fix : comments 기능 보완 완료

* feat : review 기능구현

* feat : Review 기능 구현

+++ server/src/main/java/team017/review/entity/Review.java
@@ -26,8 +26,7 @@ public class Review extends Auditable {
     private String image;
 
     @Column(nullable = false)
-    private Double star;
-
+    private int star;
     /* 💝 소비자 - 리뷰 다대일 연관 관계 : 소비자 참조 */
     @ManyToOne
     @JoinColumn(name = "client_id")
@@ -39,7 +38,7 @@ public class Review extends Auditable {
     private Board board;
 
     @Builder
-    public Review(Long reviewId, String context, String image, Double star) {
+    public Review(Long reviewId, String context, String image, int star) {
         this.reviewId = reviewId;
         this.context = context;
         this.image = image;
